+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use a cache to fill frequent requests to the Template API to reduce overall load on Lambdas
 - Enable support for development using local Lambdas
 - Enable support for organisations and organisation management through the admin panel
+- Added validation on FileInput component to ensure the type and size of the file is valid
 
 ### Changed